// parking/types/parking.ts
interface IParking {
  ADDR: string[];
  ADD_RATES: string[];
  ADD_TIME_RATE: string[];
  BUS_ADD_RATES: string[];
  BUS_ADD_TIME_RATE: string[];
  BUS_RATES: string[];
  BUS_TIME_RATE: string[];
  CAPACITY: string[];
  CUR_PARKING: string[];
  CUR_PARKING_TIME: string[];
  DAY_MAXIMUM: string[];
  FULLTIME_MONTHLY: string[];
  GRP_PARKNM: string[];
  HOLIDAY_BEGIN_TIME: string[];
  HOLIDAY_END_TIME: string[];
  HOLIDAY_PAY_NM: string[];
  HOLIDAY_PAY_YN: string[];
  LAT: string[];
  LNG: string[];
  NIGHT_FREE_OPEN: string[];
  NIGHT_FREE_OPEN_NM: string[];
  OPERATION_RULE: string[];
  OPERATION_RULE_NM: string[];
  PARKING_CODE: string[];
  PARKING_NAME: string[];
  PARKING_TYPE: string[];
  PARKING_TYPE_NM: string[];
  PAY_NM: string[];
  PAY_YN: string[];
  QUE_STATUS: string[];
  QUE_STATUS_NM: string[];
  RATES: string[];
  SATURDAY_PAY_NM: string[];
  SATURDAY_PAY_YN: string[];
  SH_CO: string[];
  SH_ETC: string[];
  SH_LINK: string[];
  SH_TYPE: string[];
  TEL: string[];
  TIME_RATE: string[];
  WEEKDAY_BEGIN_TIME: string[];
  WEEKDAY_END_TIME: string[];
  WEEKEND_BEGIN_TIME: string[];
  WEEKEND_END_TIME: string[];
}

export default IParking;